requirements.txt
bcrypt==3.1.7
celery==4.2.1
Cython==0.29.14
Django==2.2.12
django-ipware==2.1.0
django-redis==4.10.0
djangorestframework==3.11.0
flower==0.9.3
gevent==1.4.0
itsdangerous==1.1.0
Pillow==7.1.1
psycopg2-binary==2.8.4
Pygments==2.2.0
redis==3.2.0
sentry-sdk==0.4.3
tornado==5.1.1
Unidecode==1.0.22
sentry-sdk==0.4.3
uwsgi==2.0.17.1